+++ src/routes/[user]/components/Calendar/Calendar.svelte
@@ -6,8 +6,8 @@
 
   import { headerHeight, pixelsPerHour, timestampsColumnWidth } from './store.js'
   import { TOTAL_COLUMNS, COLUMN_WIDTH, c, originDT } from './constants.js'
-  import { setupCalListener, treesByDate, googleEventsByDate } from './service.js'
-  import { fetchGoogleEvents } from '$lib/utils/cloudFunctions'
+  import { setupCalListener, treesByDate } from './service.js'
+  import { getGoogleEvents } from './gcalService.js'
   import { jumpToToday } from './autoScrolling.js'
   import { trackHeight } from '$lib/utils/svelteActions.js'
   import { onMount, getContext } from 'svelte'
@@ -38,8 +38,8 @@
     const start = originDT.plus({ days: renderedLeft })
     const end = originDT.plus({ days: renderedRight })
     clearTimeout(fetchTimeout)
-    fetchTimeout = setTimeout(() => {
-       fetchGoogleEventsForRange(start, end)
+      fetchTimeout = setTimeout(() => {
+       getGoogleEvents(start, end)
     }, 500)
   }
 
@@ -57,50 +57,6 @@
     )
   })
 
-  async function fetchGoogleEventsForRange(startDT, endDT) {
-    try {
-      const timeMin = startDT.startOf('day').toISO()
-      const timeMax = endDT.endOf('day').toISO()
-      const result = await fetchGoogleEvents({ timeMin, timeMax })
-      
-      if (result.data && result.data.events) {
-        const events = result.data.events
-        const eventsByDate = {}
-        
-        for (const event of events) {
-          // Filter out cancelled events or those without dates
-          if (event.status === 'cancelled') continue
-          
-          // Only include timed events (not all-day events)
-          if (!event.start?.dateTime || !event.end?.dateTime) continue
-          
-          const startDate = event.start.dateTime.split('T')[0] // YYYY-MM-DD
-          
-          if (!eventsByDate[startDate]) eventsByDate[startDate] = []
-          eventsByDate[startDate].push({
-            id: event.id,
-            summary: event.summary,
-            start: event.start, // { dateTime: ..., timeZone: ... }
-            end: event.end,
-            colorId: event.colorId, // Google color ID
-            description: event.description,
-            location: event.location,
-            htmlLink: event.htmlLink
-          })
-        }
-
-        // Merge with existing store data
-        googleEventsByDate.update(current => {
-           // We need to be careful not to overwrite existing data if we fetch partial ranges, 
-           // but we also want to update if data changed. 
-           // Since we fetch by range, we can just spread the new data in.
-           return { ...current, ...eventsByDate }
-        })
-      }
-    } catch (e) {
-      console.error("Failed to fetch Google Events", e)
-    }
-  }
 
   function addFutureListener () {
     setupCalListener(
